requirements.txt
streamlit>=1.28.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
python-docx>=1.1.0
langdetect>=1.0.9
python-dotenv>=1.0.0
requests>=2.31.0